clean up comments, rename files, and update Postman collection
Cleaned up comments in some model files

Renamed upvote files to 'vote'

Added upvote requests to Postman collection for testing
