+++ routes/index.js
@@ -3,5 +3,5 @@ module.exports = {
   projectsRouter: require("./projects"),
   usersRouter: require("./users"),
   indexRouter: require("./home"),
-  upvotesRouter: require("./upvote"),
+  votesRouter: require("./vote"),
 };
