+++ services/index.js
@@ -3,5 +3,5 @@ module.exports = {
   ProjectService: require("./ProjectService"),
   TypeService: require("./TypeService"),
   RoleService: require("./RoleService"),
-  UpvoteService: require("./UpvoteService"),
+  VoteService: require("./VoteService"),
 };
